Adapted the script to produce new outlier fig.
Three histograms are shown instead of 4. I created obs.nz (non zero
data) which is returned from the proc.outlier function. This data is
used in the second histogram.
The 'used' variable is passed as a string and for the vertical lines we
need the position of the selected cutoff so it is extracted before
calling the Hists.